feat: roles no User para controle de acesso por feature (#36)
- roles String[] no model User (default [])
- GET /users/me retorna { id, roles } via UserMeResponseDto
- /mvp protegido por server component: roles.includes('mvp')
- /waitlist para usuários sem acesso
