+++ apps/api/prisma/schema.prisma
@@ -10,6 +10,7 @@ datasource db {
 
 model User {
   id        String    @id
+  roles     String[]  @default([])
   createdAt DateTime  @default(now())
   invoices  Invoice[]
 }
