+++ apps/api/src/users/dto/user-me-response.dto.ts
@@ -1,0 +1,10 @@
+import { User } from '@prisma/client';
+
+export class UserMeResponseDto {
+  id: string;
+  roles: string[];
+
+  static from(user: User): UserMeResponseDto {
+    return { id: user.id, roles: user.roles };
+  }
+}
