+++ apps/api/src/users/users.controller.ts
@@ -1,17 +1,31 @@
 import {
   BadRequestException,
   Controller,
+  Get,
   Headers,
   Post,
   Req,
 } from '@nestjs/common';
 import { Public } from '../auth/public.decorator';
+import { CurrentUser } from '../auth/current-user.decorator';
 import { UsersService } from './users.service';
 import { WebhookVerifier } from './webhook-verifier';
+import { UserMeResponseDto } from './dto/user-me-response.dto';
 
 interface ClerkWebhookEvent {
   type: string;
   data: { id: string };
+}
+
+@Controller('users')
+export class UserProfileController {
+  constructor(private readonly usersService: UsersService) {}
+
+  @Get('me')
+  async getMe(@CurrentUser() userId: string): Promise<UserMeResponseDto> {
+    const user = await this.usersService.getMe(userId);
+    return UserMeResponseDto.from(user);
+  }
 }
 
 @Controller('webhooks')
