+++ apps/api/src/users/users.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { UsersController } from './users.controller';
+import { UsersController, UserProfileController } from './users.controller';
 import { UsersService } from './users.service';
 import { UsersRepository } from './users.repository';
 import { WebhookVerifier } from './webhook-verifier';
 
 @Module({
-  controllers: [UsersController],
+  controllers: [UsersController, UserProfileController],
   providers: [UsersService, UsersRepository, WebhookVerifier],
 })
 export class UsersModule {}
