+++ apps/api/src/users/users.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { User } from '@prisma/client';
 import { UsersRepository } from './users.repository';
 
 @Injectable()
@@ -11,5 +12,11 @@ export class UsersService {
 
   async handleUserDeleted(id: string): Promise<void> {
     await this.repository.delete(id);
+  }
+
+  async getMe(userId: string): Promise<User> {
+    const user = await this.repository.findById(userId);
+    if (!user) throw new NotFoundException(`User ${userId} not found`);
+    return user;
   }
 }
